Move input cloning to passes
